ref/moved ImageStorageService
